Module_12/requirements.txt
confluent-kafka==2.2.0
pymongo==4.6.1


opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-zipkin
opentelemetry-instrumentation
